sql view is working

+++ aut11/database/migrations/2025_02_08_063418_example_view.php
@@ -4,26 +4,34 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Validation\Rules\Exists;
 
 return new class extends Migration
 {
     /**
      * Run the migrations.
      */
-    public function create(): void
+    public function up(): void
     {
-            DB::statement("
+        $viewName = 'userSelect';
+
+        $exists = DB::select("SELECT TABLE_NAME FROM information_schema.views WHERE TABLE_SCHEMA = DATABASE() AND TABLE_NAME = ?", [$viewName]);
+
+        if (!empty($exists)) {
+            DB::statement("DROP VIEW userSelect");
+        }
+        
+        DB::statement("
             CREATE VIEW userSelect AS
             SELECT * FROM `users` WHERE role=1;
         ");
-            
     }
 
     /**
      * Reverse the migrations.
      */
-    public function drop(): void
+    public function down(): void
     {
-        DB::statement("DROP VIEW userSelect");
+        /* DB::statement("DROP VIEW IF EXISTS userSelect"); */
     }
 };
